+++ src/client/decorators/host.ts
@@ -1,0 +1,12 @@
+import { PlusElement } from '../../types';
+import { defineProperty, host } from '../utils/index.js';
+
+export function Host() {
+  return function (target: PlusElement, propertyKey: PropertyKey) {
+    defineProperty(target, propertyKey, {
+      get() {
+        return host(this);
+      }
+    });
+  };
+}