+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.4] - 2023-05-06
+
+### Added
+
+- Added the `Host` decorator.
+- Added the `Query` decorator.
+- Added the `QueryAll` decorator.
+
+### Removed
+
+- Removed the `query` and `queryAll` helpers.
+
 ## [0.7.3] - 2023-05-06
 
 ### Added